refactor: move things into domains/foobar
