+++ src/components/foobar/pages.tsx
@@ -4,17 +4,20 @@ import styled from "styled-components";
 
 import { Button } from "@/components/Button";
 import { ViewsCounter } from "@/components/ViewsCounter";
-import { initialFoobarData } from "@/components/foobar";
 import { ShowCompletedBadges } from "@/components/foobar/badges";
 import { KonamiWrapper } from "@/components/foobar/konami";
 import { SupportSreetamDas } from "@/components/foobar/styled";
 import { Terminal } from "@/components/foobar/terminal";
 import { DocumentHead } from "@/components/shared/seo";
-import { useFoobarStore } from "@/domains/Foobar";
+import {
+	useFoobarStore,
+	TFoobarSchrodingerProps,
+	TFoobarPage,
+	initialFoobarData,
+} from "@/domains/Foobar";
 import { StyledPre } from "@/styles/blog";
 import { Space, Center } from "@/styles/layouts";
 import { Title } from "@/styles/typography";
-import { TFoobarSchrodingerProps, TFoobarPage } from "@/typings/console";
 import { dog } from "@/utils/console";
 import Custom404 from "pages/404";
